+++ tests/test_AxesStereo.py
@@ -77,6 +77,37 @@ def test_AxesStereo2D():
     axstereo.set_title('title')
     assert True
 
+def test_AxesStereo2D_zlim():
+    x = y = z = np.arange(10)
+    axstereo = AxesStereo2D()
+    axstereo.plot(x, y, z/2 + 1)
+    assert axstereo.z_lim == (1, 5.5)
+
+    axstereo.plot(x, y, z)
+    assert axstereo.z_lim == (0, 9)
+
+    axstereo.plot(x, y, z/2)
+    assert axstereo.z_lim == (0, 9)
+
+    axstereo.set_zlim((0, 5), z_autoscale=True)
+    assert axstereo.z_lim == (0, 5)
+
+    axstereo.redraw()  # we set autoscale=True, so redraw() should reset z_lim
+    assert axstereo.z_lim == (0, 9)
+
+    axstereo.set_zlim((0, 5), z_autoscale=False)
+    axstereo.plot(x, y, z)
+    assert axstereo.z_lim == (0, 5)
+
+    axstereo.plot(x, y, z, z_lim=(1, 2))
+    assert axstereo.z_lim == (1, 2)
+
+    assert axstereo.get_zlim() == (1, 2)
+
+    n_artists = 5  # from above plot()'s
+    assert len(axstereo.artists_left) == n_artists
+    assert len(axstereo.artists_right) == n_artists
+    assert len(axstereo.artist_args) == n_artists
 
 def test_AxesStereo3D():
     # Smoke test plotting
@@ -172,7 +203,25 @@ def plotting_tests_2d_pairwise():
     axstereo.stem(x, y, z, 'k')
 
 
-def plotting_tests_2d_pairwise_z_zero():
+def plotting_tests_2d_pairwise_zlim():
+    # test plot and scatter
+    x, y, z = _testdata()['trefoil']
+    axstereo = AxesStereo2D(eye_balance=-1.0)
+    axstereo.plot(x, y, z, c='k', alpha=0.2)
+    axstereo.scatter(x, y, z, c=z, cmap='viridis', s=10)
+    axstereo.grid(True)
+    axstereo.set_zlim((-0.2, 0.2))
+    axstereo.set_title('zlim=(-0.2, 0.2)')
+
+    axstereo = AxesStereo2D(eye_balance=1.0)
+    axstereo.plot(x, y, z, c='k', alpha=0.2)
+    axstereo.scatter(x, y, z, c=z, cmap='viridis', s=10)
+    axstereo.grid(True)
+    axstereo.set_zlim((-2, 2))
+    axstereo.set_title('zlim=(-2, 2)')
+
+
+def plotting_tests_anaglyph_pairwise_z_zero():
     # test plot and scatter
     x, y, z = _testdata()['trefoil']
     axstereo = AxesAnaglyph(z_zero=min(z))
@@ -220,8 +269,9 @@ def plotting_tests_anaglyph_imshow_stereo():
 
 if __name__ == '__main__':
     plotting_tests_2d_pairwise()
+    plotting_tests_2d_pairwise_zlim()
     plotting_tests_3d()
     plotting_tests_anaglyph_pairwise()
-    plotting_tests_2d_pairwise_z_zero()
+    plotting_tests_anaglyph_pairwise_z_zero()
     plotting_tests_anaglyph_imshow_stereo()
     plt.show()